# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/CountryDTO.java
package ar.edu.utn.frc.tup.lciii.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
@Data
public class CountryDTO {
    private String code;
    private String name;
}